PI-294 initial refactor of user things into the cperkins app
