+++ stackdio/users/serializers.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2014,  Digital Reasoning
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
+from django.contrib.auth import get_user_model
+from rest_framework import serializers
+
+from . import models
+
+
+class UserSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = get_user_model()
+        lookup_field = 'username'
+        fields = (
+            'url',
+            'username',
+            'first_name',
+            'last_name',
+            'email',
+            'last_login'
+        )
+
+
+class UserSettingsSerializer(serializers.HyperlinkedModelSerializer):
+    user = UserSerializer(read_only=True)
+
+    class Meta:
+        model = models.UserSettings
+        lookup_field = 'username'
+        fields = (
+            'user',
+            'public_key',
+        )
